+++ src/ladb_opencutlist/ruby/constants.rb
@@ -1,7 +1,7 @@
 module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
-    EXTENSION_VERSION = '5.1.0'.freeze
-    EXTENSION_BUILD = '202310180835'.freeze
+    EXTENSION_VERSION = '5.1.1'.freeze
+    EXTENSION_BUILD = '202310200807'.freeze
 
 end